third_party/blink/web_tests/platform/generic/fast/css/stylerule-set-selectortext-ns-expected.txt

Should not be allowed to set selectorText with unknown namespace

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(p).color is green
PASS sheet.cssRules[1].selectorText is ".nomatch"
PASS getComputedStyle(p).color is green
PASS successfullyParsed is true

TEST COMPLETE


This text should be green
